[FLINK-9816][netty] add statically linked netty-tcnative
With one of the artifacts of netty-tcnative [1], this enables support for using
the OpenSSL engine instead of the one from the JDK. [2] claims significant
improvements (3x) over the JDK and we may just also save some resources.

From the available artifacts, this commits selects the one which is statically
linked against Google's BoringSSL. Please refer to [1] for further details.

Due to possible licensing issues with the openSSL license BoringSSL is based on,
we only build this module if the build profile
'include-netty-tcnative-static-Pinclude-netty-tcnative-static' is activated.

[1] http://netty.io/wiki/forked-tomcat-native.html
[2] https://netty.io/wiki/requirements-for-4.x.html#wiki-h4-4